separate out functions , refactor
